LayoutTests/platform/glib/imported/w3c/web-platform-tests/focus/focus-restoration-in-different-site-iframes-expected.txt


FAIL Check result assert_equals: Check log expected "outerlog:log:willfocusinput,bodyfocus,inputfocus,didfocusinput,inputblur,bodyblur,bodyfocus,inputfocus," but got "outerlog:log:willfocusinput,didfocusinput,"
